moved all node functions to a seperate file for better readability
