+++ node.py
@@ -1,258 +1,10 @@
-import os
-import copy
 import torch
-from torchvision.transforms import ToTensor
 import numpy as np
 from PIL import Image
-import logging
-from torch.hub import download_url_to_file
-from urllib.parse import urlparse
 import folder_paths
 import comfy.model_management
-from .sam_hq.predictor import SamPredictorHQ
-from .sam_hq.build_sam_hq import sam_model_registry
-from .local_groundingdino.datasets import transforms as T
-from .local_groundingdino.util.utils import clean_state_dict as local_groundingdino_clean_state_dict
-from .local_groundingdino.util.slconfig import SLConfig as local_groundingdino_SLConfig
-from .local_groundingdino.models import build_model as local_groundingdino_build_model
+from .node_funtions import *
 
-logger = logging.getLogger('comfyui_segment_anything')
-to_tensor = ToTensor()
-
-sam_model_dir = os.path.join(folder_paths.models_dir, "sams")
-sam_model_list = {
-    "sam_vit_h (2.56GB)": {
-        "model_url": "https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth"
-    },
-    "sam_vit_l (1.25GB)": {
-        "model_url": "https://dl.fbaipublicfiles.com/segment_anything/sam_vit_l_0b3195.pth"
-    },
-    "sam_vit_b (375MB)": {
-        "model_url": "https://dl.fbaipublicfiles.com/segment_anything/sam_vit_b_01ec64.pth"
-    },
-    "sam_hq_vit_h (2.57GB)": {
-        "model_url": "https://huggingface.co/lkeab/hq-sam/resolve/main/sam_hq_vit_h.pth"
-    },
-    "sam_hq_vit_l (1.25GB)": {
-        "model_url": "https://huggingface.co/lkeab/hq-sam/resolve/main/sam_hq_vit_l.pth"
-    },
-    "sam_hq_vit_b (379MB)": {
-        "model_url": "https://huggingface.co/lkeab/hq-sam/resolve/main/sam_hq_vit_b.pth"
-    },
-    "mobile_sam(39MB)": {
-        "model_url": "https://github.com/ChaoningZhang/MobileSAM/blob/master/weights/mobile_sam.pt"
-    }
-}
-
-groundingdino_model_dir = os.path.join(
-    folder_paths.models_dir, "grounding-dino")
-groundingdino_model_list = {
-    "GroundingDINO_SwinT_OGC (694MB)": {
-        "config_url": "https://huggingface.co/ShilongLiu/GroundingDINO/resolve/main/GroundingDINO_SwinT_OGC.cfg.py",
-        "model_url": "https://huggingface.co/ShilongLiu/GroundingDINO/resolve/main/groundingdino_swint_ogc.pth",
-    },
-    "GroundingDINO_SwinB (938MB)": {
-        "config_url": "https://huggingface.co/ShilongLiu/GroundingDINO/resolve/main/GroundingDINO_SwinB.cfg.py",
-        "model_url": "https://huggingface.co/ShilongLiu/GroundingDINO/resolve/main/groundingdino_swinb_cogcoor.pth"
-    },
-}
-
-
-def list_files(dirpath, extensions=[]):
-    return [f for f in os.listdir(dirpath) if os.path.isfile(os.path.join(dirpath, f)) and f.split('.')[-1] in extensions]
-
-
-def list_sam_model():
-    return list(sam_model_list.keys())
-
-
-def load_sam_model(model_name ):
-    sam_checkpoint_path = get_local_filepath(
-        sam_model_list[model_name]["model_url"], sam_model_dir)
-    model_file_name = os.path.basename(sam_checkpoint_path)
-    model_type = model_file_name.split('.')[0]
-    if 'hq' not in model_type and 'mobile' not in model_type:
-        model_type = '_'.join(model_type.split('_')[:-1])
-    sam = sam_model_registry[model_type](checkpoint=sam_checkpoint_path)
-    sam.model_name = model_file_name
-    return sam
-
-
-def get_local_filepath(url, dirname, local_file_name=None):
-    if not os.path.exists(dirname):
-        os.makedirs(dirname)
-    if not local_file_name:
-        parsed_url = urlparse(url)
-        local_file_name = os.path.basename(parsed_url.path)
-    destination = os.path.join(dirname, local_file_name)
-    if not os.path.exists(destination):
-        logging.warn(f'downloading {url} to {destination}')
-        download_url_to_file(url, destination)
-    return destination
-
-
-def load_groundingdino_model(model_name, use_cpu=False):
-    dino_model_args = local_groundingdino_SLConfig.fromfile(
-        get_local_filepath(
-            groundingdino_model_list[model_name]["config_url"],
-            groundingdino_model_dir
-        ),
-
-    )
-    dino = local_groundingdino_build_model(dino_model_args)
-    checkpoint = torch.load(
-        get_local_filepath(
-            groundingdino_model_list[model_name]["model_url"],
-            groundingdino_model_dir,
-        ),
-    )
-    dino.load_state_dict(local_groundingdino_clean_state_dict(
-        checkpoint['model']), strict=False)
-    return dino
-
-
-def list_groundingdino_model():
-    return list(groundingdino_model_list.keys())
-
-
-def groundingdino_predict(
-    dino_model,
-    image,
-    prompt,
-    box_threshold,
-    device
-):
-    print(f"\033[1;32m(segment-anything) groundingdino_predict using:\033[0m {device}")
-    def load_dino_image(image_pil):
-        transform = T.Compose(
-            [
-                T.RandomResize([800], max_size=1333),
-                T.ToTensor(),
-                T.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225]),
-            ]
-        )
-        image, _ = transform(image_pil, None)  # 3, h, w
-        return image
-
-    def get_grounding_output(model, image, caption, box_threshold, device):
-        print(f"\033[1;32m(segment-anything) get_grounding_output using:\033[0m {device}")
-        caption = caption.lower()
-        caption = caption.strip()
-        if not caption.endswith("."):
-            caption = caption + "."
-        image = image.to(device)
-        with torch.no_grad():
-            outputs = model(image[None], captions=[caption])
-        logits = outputs["pred_logits"].sigmoid()[0]  # (nq, 256)
-        boxes = outputs["pred_boxes"][0]  # (nq, 4)
-        # filter output
-        logits_filt = logits.clone()
-        boxes_filt = boxes.clone()
-        filt_mask = logits_filt.max(dim=1)[0] > box_threshold
-        logits_filt = logits_filt[filt_mask]  # num_filt, 256
-        boxes_filt = boxes_filt[filt_mask]  # num_filt, 4
-        return boxes_filt.to(device)
-
-    dino_image = load_dino_image(image.convert("RGB"))
-    boxes_filt = get_grounding_output(
-        dino_model, dino_image, prompt, box_threshold, device
-    )
-    H, W = image.size[1], image.size[0]
-    for i in range(boxes_filt.size(0)):
-        boxes_filt[i] = boxes_filt[i] * torch.Tensor([W, H, W, H]).to(device)
-        boxes_filt[i][:2] -= boxes_filt[i][2:] / 2
-        boxes_filt[i][2:] += boxes_filt[i][:2]
-    return boxes_filt.to(device)
-
-
-def create_pil_output(image_np, masks, boxes_filt):
-    output_masks, output_images = [], []
-    boxes_filt = boxes_filt.numpy().astype(int) if boxes_filt is not None else None
-    for mask in masks:
-        output_masks.append(Image.fromarray(np.any(mask, axis=0)))
-        image_np_copy = copy.deepcopy(image_np)
-        image_np_copy[~np.any(mask, axis=0)] = np.array([0, 0, 0, 0])
-        output_images.append(Image.fromarray(image_np_copy))
-    return output_images, output_masks
-
-
-def create_tensor_output(image_np, masks, boxes_filt,device):
-    #device = comfy.model_management.get_torch_device()
-    output_masks, output_images = [], []
-    boxes_filt = boxes_filt.cpu().numpy().astype(int) if boxes_filt is not None else None
-    for mask in masks:
-        image_np_copy = copy.deepcopy(image_np)
-        image_np_copy[~np.any(mask, axis=0)] = np.array([0, 0, 0, 0])
-        output_image, output_mask = split_image_mask(Image.fromarray(image_np_copy),device)
-        output_masks.append(output_mask)
-        output_images.append(output_image)
-    return (output_images, output_masks)
-
-
-def split_image_mask(image, device):
-    #device = comfy.model_management.get_torch_device()
-    image_rgb = image.convert("RGB")
-    image_rgb = np.array(image_rgb).astype(np.float32) / 255.0
-    image_rgb = torch.from_numpy(image_rgb)[None,]
-    if 'A' in image.getbands():
-        mask = np.array(image.getchannel('A')).astype(np.float32) / 255.0
-        mask = torch.from_numpy(mask)[None,]
-    else:
-        mask = torch.zeros((64, 64), dtype=torch.float32, device=device)
-    return (image_rgb, mask)
-
-
-def sam_segment(
-    sam_model,
-    image,
-    boxes,
-    multimask,
-    device
-):  
-    #device = comfy.model_management.get_torch_device()
-    print(f"\033[1;32m(segment-anything) sam_segment using:\033[0m {device}")
-    if boxes.shape[0] == 0:
-        return None
-    sam_is_hq = False
-    # TODO: more elegant
-    if hasattr(sam_model, 'model_name') and 'hq' in sam_model.model_name:
-        sam_is_hq = True
-    predictor = SamPredictorHQ(sam_model, sam_is_hq)
-    image_np = np.array(image)
-    image_np_rgb = image_np[..., :3]
-    predictor.set_image(image_np_rgb)
-    transformed_boxes = predictor.transform.apply_boxes_torch(
-        boxes, image_np.shape[:2])
-    masks, _, _ = predictor.predict_torch(
-        point_coords=None,
-        point_labels=None,
-        boxes=transformed_boxes.to(device),
-        multimask_output=False)
-    
-    if multimask is not False:
-        print(f"\033[1;32m(segment-anything) sam_segment using multimask:\033[0m {multimask}")
-
-        output_images, output_masks = [], []
-        for batch_index in range(masks.size(0)):
-            mask_np =  masks[batch_index].permute( 1, 2, 0).cpu().numpy()# H.W.C
-            image_with_alpha = Image.fromarray(np.concatenate((image_np_rgb, mask_np * 255), axis=2).astype(np.uint8), 'RGBA')
-            _, msk = split_image_mask(image_with_alpha,device)
-            r, g, b, a = image_with_alpha.split()
-
-            black_image = Image.new("RGB", image.size, (0, 0, 0))
-            black_image.paste(image_with_alpha, mask=image_with_alpha.split()[3])
-
-            rgb_ts = to_tensor(black_image)
-            rgb_ts = rgb_ts.unsqueeze(0)
-            rgb_ts = rgb_ts.permute(0, 2, 3, 1)
-
-            output_images.append(rgb_ts)
-            output_masks.append(msk)
-                        
-        return (output_images, output_masks)
-    else:
-        masks = masks.permute(1, 0, 2, 3).cpu().numpy()
-        return create_tensor_output(image_np, masks, boxes, device)
 
 
 class SAMModelLoader:
@@ -284,7 +36,7 @@ class GroundingDinoModelLoader:
     FUNCTION = "main"
     RETURN_TYPES = ("GROUNDING_DINO_MODEL", )
 
-    def main(self, model_name, use_cpu=False):
+    def main(self, model_name):
         dino_model = load_groundingdino_model(model_name)
         return (dino_model, )
 
@@ -306,13 +58,14 @@ class GroundingDinoSAMSegment:
                 }),
                 "multimask": ('BOOLEAN', {"default":False}),
                 "dedicated_device": (["Auto", "CPU", "GPU"], ),
+                "optimize_prompt_for_dino":('BOOLEAN', {"default":False})
             },
         }
     CATEGORY = "segment_anything"
     FUNCTION = "main"
     RETURN_TYPES = ("IMAGE", "MASK")
 
-    def main(self, grounding_dino_model, sam_model, image, prompt, box_threshold,multimask=False, dedicated_device="Auto"):
+    def main(self, grounding_dino_model, sam_model, image, prompt, box_threshold,optimize_prompt_for_dino=False,multimask=False, dedicated_device="Auto"):
         #
         device_mapping = {
             "Auto": comfy.model_management.get_torch_device(),
@@ -321,10 +74,9 @@ class GroundingDinoSAMSegment:
         }
         device = device_mapping.get(dedicated_device)
         #
-        print(f"\033[1;32m(segment-anything) grounding_dino_model moving to:\033[0m {device}")
+        # send model to selected device 
         grounding_dino_model.to(device)
         grounding_dino_model.eval()
-        print(f"\033[1;32m(segment-anything) sam_model moving to:\033[0m {device}")
         sam_model.to(device)
         sam_model.eval()
         #
@@ -333,6 +85,7 @@ class GroundingDinoSAMSegment:
         empty_mask = torch.zeros((1, 1, img_height, img_width), dtype=torch.float32) # [B,C,H,W]
         empty_mask = empty_mask / 255.0
         #
+        # empty output
         res_images = []
         res_masks = []
         #
@@ -340,18 +93,21 @@ class GroundingDinoSAMSegment:
         #
         for item in image:
             item = Image.fromarray(np.clip(255. * item.cpu().numpy(), 0, 255).astype(np.uint8)).convert('RGBA')
+            # run dino for prompt guides segmentation
             boxes = groundingdino_predict(
                 grounding_dino_model,
                 item,
                 prompt,
                 box_threshold,
+                optimize_prompt_for_dino,
                 device
             )
 
+            # abboart if nothing is detected and set detection_errors
             if boxes.numel() == 0:
                 detection_errors = True
                 break
-
+            # create detailed masks with SAM depending on boxes found by dino
             (images, masks) = sam_segment(
                 sam_model,
                 item,
@@ -359,17 +115,21 @@ class GroundingDinoSAMSegment:
                 multimask,
                 device
             )
+            # add results to output
             res_images.extend(images)
             res_masks.extend(masks)
         
+        # if nothing was detected just send simple input image and empty mask
         if detection_errors is not False:
             print("\033[1;32m(segment-anything)\033[0m The tensor 'boxes' is empty. No elements were found in the image search.")
             res_images.append(image)
             res_masks.append(empty_mask)
 
+        # generate output
         res_images = torch.cat(res_images, dim=0)
         res_masks = torch.cat(res_masks, dim=0)
         return (res_images, res_masks, )
+
 
 class BatchSelector:
     @classmethod
@@ -396,25 +156,3 @@ class BatchSelector:
         selected_image = selected_image.unsqueeze(0)
         selected_mask = mask[selector]
         return (selected_image, selected_mask, )
-    
-
-"""
-if __name__ == "__main__":
-    input_image = Image.open(
-        '/data/dev/comfyui-latest/custom_nodes/comfyui_segment_anything/human.jpg').convert('RGBA')
-    dino_model = load_groundingdino_model('GroundingDINO_SwinT_OGC (694MB)')
-    boxes = groundingdino_predict(
-        dino_model,
-        input_image,
-        'face . glasses . forehead',
-        0.3
-    )
-    sam_model = load_sam_model('sam_hq_vit_h (2.57GB)')
-    (output_images, output_masks) = sam_segment(
-        sam_model,
-        input_image,
-        boxes
-    )
-    for i in range(len(output_images)):
-        output_images[i].save(f"result_{i}.png")
-"""
